[#143112] Fix spec when account number is not valid (#1664)
# Release Notes

Tech task: Fix spec when account number is not considered valid.

# Additional Context

In Dartmouth and possibly other schools, the "TESTING123" is not a valid format, but
for the sake of this spec, we don't care: we just need it to be considered a global
account type.

Fixes the merge of https://github.com/tablexi/nucore-open/pull/1659 into DC