[#143399] Add another column to current rooms dashboard (#1693)
# Release Notes

Adds another "Current Occupancies" column to the secure rooms dashboard so that a high occupancy (more than 10) does not overflow the screen.

# Screenshot

<img width="1252" alt="screen shot 2018-09-10 at 4 09 11 pm" src="https://user-images.githubusercontent.com/1099111/45324608-3cdeae00-b514-11e8-8793-53a14c7a9fdc.png">

# Additional Context

I chatted with Aaron and showed him a couple options. This maintains the current behavior of always showing the "Current Occupancies" header and table header of the first column for empty rooms. We hide the second column header if there is nothing there.

The facilities occupancies shows all of the rooms for a facility, but shared the same code with the public dashboard so this change applies to both places.