Use/Manage menu items (#1683)
# Release Notes

Implementation of our Use/Manage design separation. Currently, "Use" will bring the user back to the facility list home page unless they are managing a facility. If they are an admin and managing a facility, then they will be taken to the facility-specific homepage with the product list.

Manage mode will take the user to the manageable facility list, however it could also be a dropdown such as in our original design.

# Screenshot
![screen shot 2018-09-06 at 9 47 16 am](https://user-images.githubusercontent.com/5000430/45161460-e68f0980-b1b9-11e8-91c8-60cfca6bc0fd.png)