[#137154] Email user when an order goes into problem queue (#1685)
# Release Notes

Notify the user via email if their reservation/occupancy enters into the problem queue. 

# Additional Context

We do not trigger the email for missing price policies as that is something the facility needs to address.

For reservations, this only applies to timer and relay based instruments.

* Reservation never started, 12 hours after it was scheduled to end
* Reservation started, but never ended, 12 hours after it was scheduled to end
* Reservation started, but never ended and another reservation is started. Applies across a shared scheduled.
* Logged out via the Auto Logout setting (relays only)
* Reservation started and not yet ended and instrument is marked offline. Applies across a shared schedule.

For occupancies, this only applies to normal users. Facility staff does not get orders associated with their occupancies so they will never go to the problem queue.

* Scan into a room, do not scan out, 12 hours after the scan in
* Scan into a room, scan in again. First scan's order goes into the problem queue
* Scan out without an associated scan in