[#133770] Facility Notification Banner (#1703)
# Release Notes

Adds a per-facility optional banner notification that can be configured by the facility. The banner appears on the facility's home page (the instrument listing).

# Screenshot

![screen shot 2018-09-17 at 7 06 28 pm](https://user-images.githubusercontent.com/1099111/45657150-b3e4eb00-baae-11e8-9c4d-ad91b9ab8469.png)

In Dartmouth when combined with the SSO warning notice (hopefully will be going away in the next little bit).

![screen shot 2018-09-17 at 7 19 57 pm](https://user-images.githubusercontent.com/1099111/45657152-b6474500-baae-11e8-97a6-082b4591b0be.png)

Combined with the order on behalf of banner. This doesn't look great, but that banner could definitely use some love in the future. We could also hide the notification banner if the order on behalf banner.

# Additional Context

I gave this a feature flag because I could see NU not being interested.