Changes to typography (#1643)
# Release Notes
As we discussed earlier, we'll start to make some changes to spacing and typography. I'm reluctant to start making massive changes to spacing until some of the UI elements begin to change, except where improper spacing makes the site unusable.

This pull request modifies the typography of the site to fit the specification we shared.

# Screenshots

## Facilities listing
### Before
![facilities-nucore-upstream](https://user-images.githubusercontent.com/3682844/43665418-09a723e4-973e-11e8-8732-a4ccd19a934a.png)
### After
![facilities-design-rework-typography](https://user-images.githubusercontent.com/3682844/43665416-099d1cc8-973e-11e8-873d-5a01a63892f2.png)

## Example Facility
### Before
![facilities_example-nucore-upstream](https://user-images.githubusercontent.com/3682844/43665415-098f8ad6-973e-11e8-9132-7bc5298be688.png)
### After
![facilities_example-design-rework-typography](https://user-images.githubusercontent.com/3682844/43665414-0985b45c-973e-11e8-90df-55cbb2f95e7c.png)

# Additional Context
If any spacing looks egregiously incorrect, let me know and I will fix it. I'm just reluctant to make every change in a single PR.