[#136151] Ability to charge full reservation cost for cancellation (#1708)
# Release Notes

Adds the ability to charge the full reservation cost (based on reservation times) on cancellation. This take precedence over the "Reservation/Cancellation Cost"

# Screenshot

<img width="925" alt="screen shot 2018-09-20 at 4 44 42 pm" src="https://user-images.githubusercontent.com/1099111/45848547-89846f00-bcf4-11e8-84ef-4c69b5e39219.png">

# Additional Context

This could have compliance issues for NU, so that's why it's being feature flagged. Also, Dartmouth overrides the `cancellation_cost` locale from "Reservation Cost" to "Cancellation Cost" so we'll want to adjust the local for this new field to match for them.