Localize account member roles (#1728)
# Release Notes

Tech task: Added localization to account member roles.

# Screenshot

<img width="894" alt="screen shot 2018-10-09 at 9 42 00 am" src="https://user-images.githubusercontent.com/5000430/46673610-2a55a400-cba8-11e8-862c-f0688944ebcb.png">

# Additional Context

Resolves https://github.com/SquaredLabs/nucore-uconn/issues/37

Because the roles are hard coded and in the database, there is no way to elegantly allow them to be localized. 
See: `account_user.rb`
```ruby
  ACCOUNT_PURCHASER = "Purchaser"
  ACCOUNT_OWNER = "Owner"
  ACCOUNT_ADMINISTRATOR = "Business Administrator"
```

Business administrator has a space, so it cannot be used to query i18n. My best workaround is to convert it in the view:
```haml
%td= t(".role.#{au.user_role.downcase.parameterize.underscore}")
```