[#143568] Use Roboto font for PDFs (#1734)
# Release Notes

Use Roboto font when generating PDFs in order to support UTF-8 characters.

# Additional Context

This will prevent `ActionView::Template::Error: Your document includes text that's not compatible with the Windows-1252 character set. If you need full UTF-8 support, use TTF fonts instead of PDF's built-in fonts.`

I went with Roboto as it's a free and widely used font. It seemed a reasonable default replacing the Helvetica that was the PDF's default.