Add manage mode class to other/global pages (#1745)
# Release Notes

Update the logic to check which mode the page is in and fix manage mode class on pages that were missing it.

# Additional Context

It seems like the issue was not necessarily that the use/manage class was not correct in the helper. The fix ended up being that some views don't use `application.html.haml`, there are also two-column layouts that it needed to be added to.