[#144525] Alphabetize homepage case insensitively (#1752)
# Release Notes

Fix issue where facilities beginning with lowercase letters were shifted to the end of the list.

# Additional Context

Dartmouth has a lab called "bioMT" which was hard to find because it was pulled down to the bottom. 

This also fixes an issue with the A-Z list where it treats "b" and "B" as separate letters.