[#148409] Reduce number of DB queries on Facility timeline page (#1899)
* Refactor away Timelinable module

The indirection and hack (in ReservationsController, public_timeline calls timeline) here are not worth the DRY aspect, and the two pages are very different, so decoupling their implementation allows us to evolve and optimize the controller methods separately as it makes sense.

* Separate facility and public schedule partials and get rid of conditional

* Refactor association-like methods on Schedule to associations

* Use .size on association in _schedule so we don’t have to execute a count every time

* Eliminate N+1 for loading alerts when displaying timelines

* Fold single-use partial _relay_switch into only call site

* Avoid N+1 for relays on facility timeline

* Avoid two unnecessary queries just to check if an instrument is offline

The method #offline? needs to check whether there are any current offline reservations. By refactoring to an association and using #present?, Rails will memoize the association and check without another DB query.

* Appease rubocop’s style preferences