[#147850] Set ordered_at when adding to an order (#2117)
# Release Notes

Set ordered_at to "now" or match fulfilled at when adding onto an existing order.

# Additional Context

This continues from #2116.

When ordering on behalf: Orders take on whatever date/time you put it under the "More options". If left blank, or unexpanded, it defaults to setting ordered_at to "now". This behavior should be unchanged.

When adding to an order:

"New" orders - get "now" as the ordered_at. If it's a reservation/service with order form, the order will receive the ordered_at of when the details gets fully added to the order (once you've resolved any issues like uploading the order form).
"Complete" - if you set a fulfilled at, both fulfilled_at and ordered_at will take on that time (the logic we've used is 12:00 noon Chicago time of the date you put in)
If you don't set a fulfilled date, both fulfilled_at ordered_at will fall back to "now".

Imported orders have an ordered at column. This will continue to be used. It's just the location of the column is on a different table.

Secure Rooms/Doors should maintain their same behavior.