Pull nu_research_safety engine into open (#2131)
# Release Notes

Pull the `nu_research_safety` engine out of `nucore-nu` into `nucore-open`.

# Additional Context

This is a straight copy/paste of what's left in the engine after https://github.com/tablexi/nucore-nu/pull/541

I plan to in another PR to pull all of this code into the main app and put it behind a feature flag. Then the engine should go away.