Add instructions for running LDAP in docker-compose (#2115)
# Release Notes

Update instructions for running a dev LDAP server within `docker-compose`.

# Additional Context

I did this as I was working on #2108 converting external to internal users. I initially left it in the `docker-compose.yml` file commented out, but I figured it won't be used enough it's not worth giving it such a prominent spot.