Remove constricting spec (#2148)
# Release Notes

Tech task: remove spec that restricts what locations we can put uploaded files in.

# Additional Context

This spec makes it very difficult to use a different pattern for the storage location. Now
we can change the setting to be whatever we want and it won't break specs.

In UMass we're storing them in `public/uploads`.