[#152343] Add auditing to bulk account deletion (#2151)
# Release Notes

Add auditing when bulk removing accounts from a user. This auditing already exists when individually removing a user from an account.

# Additional Context

Before, soft-delete was done manually. This adds paranoia to the `AccountUser` model, but does not add a default scope since we're already hiding the soft-deleted records where we need to. In addition, it adds a `LogEvent` when accounts are removed (which we currently do in individual removal).

This also prevents bulk removal of account owner records. This was previously only done by disabling the checkbox, but now we do it in the controller as well.

Original ticket: https://pm.tablexi.com/issues/144321

Relevant PRs: #1917 #1919 #1922 #1927