Fix recharge account editing (#2165)
# Release Notes

Fix recharge account editing. Clicking "Save" would take the user to a 404 page.

# Additional Context

Editing a recharge account was broken and we didn't have a good spec around it. It's possible it was broken in #2136 by changing `form_for` to `simple_form_for`.

The spec is a little funky in how it fills out the form, but it needs to be in order for NU and Dartmouth to pass since they use different sets of fields.