[#146482] Users should not be able to resolve if missing start time (#2174)
# Release Notes

Change the behavior of the new problem resolution feature so that a user may not resolve their own reservation if they did not start the reservation.

# Additional Context

First introduced in #2119, a user can now fix their own problem reservation if they forgot to log off. They should not however, be able to fix it if they forgot to log on as well. The feature as built doesn't support editing the start time.

A reservation may end up in this state if the User/Manager creates a reservation and never starts the reservation. 12 hours after the reservation end time, it will move to the Problem Queue missing both start and end times.

This also fixes a UI issue where choosing "today" displays an error of "cannot be before M/DD/YYYY".