Fix downstream specs (#2181)
# Release Notes

Tech task: Fix specs that break downstream in the schools.

# Additional Context

I've run these changes in all three schools, but I'm only issuing the PR here for now.

The second commit is my first attempt at fixing the recharge feature spec. It's particularly nasty, though. It needs to account for making sure Dartmouth doesn't make an API call, create the `api_speed_type` for UMass (which we only do on `create`, and handle a case where NU is setting the components of a chart string lazily so trying to change an attribute of a destroyed model triggers an error.

I ended up deciding it's much cleaner to conditionally run the spec based on if we're in `open` or a school that hasn't built custom validation yet.