Fix caching issue with shared relays (#2213)
# Release Notes

Fix issue where second relay on a shared schedule would be left in a spinning state on the timeline view.

# Additional Context

Before, in the response we would get two entries but they'd be sharing the same `instrument_id`. This was causing the second relay to be left in a spinning state. This now makes sure that if there are two relays on different instruments, we return different `instrument_id`s.