Switch from Poltergeist/PhantomJS to Chrome headless and Rails system tests (#2214)
# Release Notes

Tech tasks: Switch to using Rails system specs. Switch from using Poltergeist/PhantomJS to Chrome Headless as the former is no longer maintained.

# Additional Context

I found this as I was working on some newer JS which was breaking in PhantomJS. Switching to system specs is a good thing to do anyways now that we're at Rails 5.2 since they claim to be faster.

A lot of the changes were simply renaming  files from `spec/features` to `spec/system`.