[#151885] Fixes for add to order on bulk upload (#2220)
# Release Notes

Fixes for adding to an existing order in bulk upload.

# Additional Context

Fixes to #2167.

> If the Payment Source varies for order details that are being added to an order, the Payment Source displayed in NUcore will only reflect the first Payment Source that is designated in the upload file. Behavior should be that the Payment Source matches the upload (assuming the Payment Source is valid, and the Purchaser is on it).

What was happening is for adding-on lines, it was taking the original order's account. It was still validating that user has permission to use the chart string specified in the upload; it just wasn't persisting it.

> If a Bulk Upload contains some rows with an Order number and other rows without an Order number, all orders will still be added to the Order. Behavior should be that lines without an Order should generate new Order ID's.

The specific case is: a line item that adds to an order; a line item that does not add to an order, but matches the user, chartstring, and order date of the first line. If they didn't match, they would get split out onto separate orders.

The way the upload always (well, at least the past long while) behaved before this feature and how it behaves now for non-add-to orders (with the column blank): We roll up all the lines that have the same user, chartstring, and order date and put them onto the same order. We essentially build a cache of orders as we go down the CSV and key them by those fields.

In the new feature, if the cache got populated with an existing order, future orders that match the keys would get added to that existing order, even if the order number field was blank. This also was a problem if the order number was different, it would end up on the first specified order.