[#152849] Fix movable transaction when order is missing costs (#2221)
# Release Notes

Fix for movable transaction when order is missing costs.

# Additional Context

This would usually be a problem order, or could also be a canceled order.

There are two errors `NoMethodError: undefined method '+' for nil:NilClass` and `TypeError: nil can't be coerced into BigDecimal`. They're both manifestations of the same thing, just depends on the order: `nil + something` or `something + nil`.