[#137160] Reorganize Billing Sidebar and remove AR module (#2211)
# Release Notes

Reorganize the billing sidebar into a more linear progression of steps. Remove the accounts receivable module in favor of improved functionality in "Statement History".

# Additional Context

From NU:

> There are multiple issues with the AR tab, and now that there's all this great functionality in "Statement History" it really doesn't serve a useful purpose. 
>
> The Show/Hide link in AR never seemed to work, and AR for internal chart strings isn't super helpful if the cores are billing regularly in any case. Statement History is a much better AR module.

This also removes the `manage_payment_sources_with_users` setting. Way back when, the "Payment Sources" tab was in the billing tab. We moved it to the Users tab for NU, but left it where it was for UIC. Now it will be in Users for everyone.

# Screenshot

### Before

![Screen Shot 2020-02-13 at 1 41 51 PM](https://user-images.githubusercontent.com/1099111/74471898-9e701e00-4e66-11ea-86fa-f2c213aea8ce.png)

### After

![Screen Shot 2020-02-13 at 1 41 37 PM](https://user-images.githubusercontent.com/1099111/74471910-a465ff00-4e66-11ea-9583-0b7d9684408a.png)