[#140079] Improve product search (MySQL only) (#2209)
# Release Notes

Improve product search responses, especially when searching for multiple words. For now, this only applies to MySQL.

For example: searching for "lsc W1 spinning disk" might return "LSC Andor Spinning Disk" and "LSC Nikon A1RSi" where it's matches a single word. Precedence is given to the result that matches more words. Before it would return no results.

# Additional Context

This uses MySQL's full text searching functionality.

It still may not be perfect, especially for things like misspellings, or word roots, but it's the best we can do without a fuller search engine, which would likely require additional infrastructure (e.g. Sphinx).

I have code that works for Oracle as well, but there are problems with the oracle adapter currently. I have PRs out for it, but I don't know how long that will take to go in. In the meantime, I'll open a PR here for it so you can see it. Some of the code in this PR may feel a little over-flexible, but it made sense with the Oracle support.