[#151928] Fix hard error on cross-facility Payment Sources (#2243)
# Release Notes

Fixes a hard error on global "Users > Payment Sources". (`/facilities/all/accounts`).

# Additional Context

This was broken by the removal of the `accountsfacility_id` column in #1863. We didn't have a test around that view, which would have caught it. This changes the behavior slightly: before we were only showing global accounts before. But when you searched that restriction when away. Now on the page, it'll include credit card and POs. I also added a facility column that will appear in the cross-facility context.

# Screenshot

![Screen Shot 2020-03-06 at 2 13 10 PM](https://user-images.githubusercontent.com/1099111/76119228-9eba8f80-5fb4-11ea-8028-2f9c992e46ad.png)