[#149739] Fix sorting when deciding most recent order (#2246)
# Release Notes

Fix "Add to Order" issue where the defaults may be coming from the wrong order detail.

# Additional Context

See #2207 for the original implementation

In Oracle, it's possible that the IDs of the rows in `order_details` (or really in any table) may not be assigned in the order of the sequence. This is due to the way that Oracle caches values in a sequence: it preallocates them into a cache, but different nodes get their own cache. So you can end up with an `ORDER_BY id` being different than `ORDER BY created_at`.

In the future, we should consider how to ensure proper ordering (see the references), but that is out of scope of the current ticket.

A couple references:
https://asktom.oracle.com/pls/apex/f?p=100:11:::NO::P11_QUESTION_ID:9539374000346423410
https://stackoverflow.com/questions/13513936/oracle-sequence-value-are-not-ordered
https://stackoverflow.com/questions/4866959/oracle-rac-and-sequences